data loader uses existing setup functions
